% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{add_to_mda}
\alias{add_to_mda}
\title{Title}
\usage{
add_to_mda(
  mda_info,
  mda_start_time,
  last_mda_time,
  regularity,
  drug_efficacy,
  pre_SAC_prop,
  SAC_prop,
  adult_prop
)
}
\arguments{
\item{adult_prop}{}
}
\value{

}
\description{
Title
}
