% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{administer_drug}
\alias{administer_drug}
\title{Administer drugs to specified individuals.}
\usage{
administer_drug(humans, indices, drug_effectiveness)
}
\arguments{
\item{humans}{human population}

\item{indices}{which individuals to administer drug to}

\item{drug_effectiveness}{how effective the drug is}
}
\description{
Administer drugs to specified individuals.
}
