% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{calculate_worm_pairs}
\alias{calculate_worm_pairs}
\title{function to calculate the number of worm pairs}
\usage{
calculate_worm_pairs(female_worms, male_worms)
}
\arguments{
\item{female_worms}{number of female worms for each individual}

\item{male_worms}{number of male worms for each individual}
}
\description{
function to calculate the number of worm pairs
}
