% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{create_contact_settings}
\alias{create_contact_settings}
\title{Create contact by age rates given a scenario}
\usage{
create_contact_settings(scenario)
}
\arguments{
\item{scenario}{one of "low adult", "moderate adult" and "high adult"}
}
\description{
Create contact by age rates given a scenario
}
