% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{create_mda}
\alias{create_mda}
\title{create the MDA}
\usage{
create_mda(
  pre_SAC_prop,
  SAC_prop,
  adult_prop,
  first_mda_time,
  last_mda_time,
  regularity,
  pre_SAC_gender,
  SAC_gender,
  adult_gender,
  mda_effectiveness
)
}
\arguments{
\item{pre_SAC_prop}{proportion of pre-SAC individuals to recieve MDA}

\item{SAC_prop}{proportion of SAC individuals to recieve MDA}

\item{adult_prop}{proportion of adults individuals to recieve MDA}

\item{first_mda_time}{specifies when this will first occur in years}

\item{last_mda_time}{specifies when this will last occur in years}

\item{regularity}{how regularly the mda will repeat in years}

\item{pre_SAC_gender}{gender of pre-SAC individuals to recieve MDA}

\item{SAC_gender}{gender of SAC individuals to recieve MDA}

\item{adult_gender}{gender of adults individuals to recieve MDA}

\item{mda_effectiveness}{efficacy of drug used.}
}
\description{
create the MDA
}
