% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{create_population_specified_ages}
\alias{create_population_specified_ages}
\title{Create the population to begin from}
\usage{
create_population_specified_ages(pars)
}
\arguments{
\item{pars}{parameters input}
}
\value{
list(humans, miracidia, cercariae)
}
\description{
Create the population to begin from
}
