% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{default_pars}
\alias{default_pars}
\title{Return the default parameters defined in the julia package}
\usage{
default_pars()
}
\value{
parameters Julia structure
}
\description{
Return the default parameters defined in the julia package
}
