% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{generate_ages_and_deaths}
\alias{generate_ages_and_deaths}
\title{Will age the human population in order to get a realistic age structure based on input death rates by age.}
\usage{
generate_ages_and_deaths(num_steps, humans, pars)
}
\arguments{
\item{num_steps}{how many steps forward to take}

\item{humans}{the humans}

\item{pars}{parameters used}
}
\value{
will return the human population container
}
\description{
Will age the human population in order to get a realistic age structure based on input death rates by age.
}
