% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{get_adult_data_from_record}
\alias{get_adult_data_from_record}
\title{output information for burden and heavy-intensity burden in adults from julia recorded data}
\usage{
get_adult_data_from_record(record)
}
\arguments{
\item{record}{variable which stores the data}
}
\description{
output information for burden and heavy-intensity burden in adults from julia recorded data
}
