% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{get_data_from_record}
\alias{get_data_from_record}
\title{output information from julia recorded data}
\usage{
get_data_from_record(record)
}
\arguments{
\item{record}{variable which stores the data}
}
\description{
output information from julia recorded data
}
