% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{get_dot_mean}
\alias{get_dot_mean}
\title{Calculate the element wise mean of whatever Julia array is input}
\usage{
get_dot_mean(a)
}
\arguments{
\item{a}{- input object to take element wise mean of}
}
\description{
Calculate the element wise mean of whatever Julia array is input
}
