% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{get_selected_data}
\alias{get_selected_data}
\title{Get all entries of a chosen variable from the Julia humans array}
\usage{
get_selected_data(humans, name)
}
\arguments{
\item{humans}{Julia array containing information about human population}

\item{name}{name of variable we want to return.
Must match exactly a name of a variable e.g. "age", "eggs", "female_worms", "male_worms"}
}
\value{
Will return an r array
}
\description{
Get all entries of a chosen variable from the Julia humans array
}
