% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{make_age_contact_rate_array}
\alias{make_age_contact_rate_array}
\title{Make the contact by age array given the specified inputs}
\usage{
make_age_contact_rate_array(pars, scenario, input_ages, input_contact_rates)
}
\arguments{
\item{scenario}{can be one of "low adult", "moderate adult" and "high adult"}

\item{input_ages}{if scenario not specified, we can input ages at which we wish contact rates to change. Ages must be integers, and input with as.integer()}

\item{input_contact_rates}{the contact rates corresponding to the ages in the input_ages variable}

\item{max_age}{maximum age for a human}
}
\value{
will return the parameters object
}
\description{
Make the contact by age array given the specified inputs
}
\examples{
\donttest{
# define input ages
input_ages = array(data = c(as.integer(4), as.integer(9), as.integer(15), as.integer(120)))

# define contact rates for age groups
input_contact_rates = array(data = c(0.64, 0.91, 1, 0.018))

# normalize so that the sum of the contact rates is 1
input_contact_rates = input_contact_rates/sum(input_contact_rates)

# make the contact by age array within the parameters object
#pars = default_pars()
#pars = make_age_contact_rate_array(pars, scenario, input_ages, input_contact_rates)
}
}
