% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{parameter_value}
\alias{parameter_value}
\title{Get the value of a chosen parameter}
\usage{
parameter_value(pars, name)
}
\arguments{
\item{pars}{the parameter object currently in use}

\item{name}{name of variable we want to return.
Must match exactly a name of a variable in parameters e.g. "contact_rate", "N" etc. Set ?set_pars for list of names}
}
\value{

}
\description{
Get the value of a chosen parameter
}
