% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{run_repeated_sims_no_population_change_increasing}
\alias{run_repeated_sims_no_population_change_increasing}
\title{run simulations where the population doesn't change and the egg production function is monotonically increasing}
\usage{
run_repeated_sims_no_population_change_increasing(
  filename,
  num_time_steps,
  mda_info,
  vaccine_info,
  num_repeats
)
}
\arguments{
\item{filename}{name of file to store population in}

\item{num_time_steps}{how many time steps to step forward}

\item{mda_info}{container for all MDAs to take place}

\item{vaccine_info}{container for all vaccination programs to take place}

\item{num_repeats}{how many times do we repeat the simulation}
}
\description{
run simulations where the population doesn't change and the egg production function is monotonically increasing
}
