% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{save_population_to_file}
\alias{save_population_to_file}
\title{Save humans, environmental cercariae and mircacidia to a file along with the parameters used.}
\usage{
save_population_to_file(filename, humans, miracidia, cercariae, pars)
}
\arguments{
\item{filename}{file to save to}

\item{humans}{humans variable}

\item{miracidia}{environmental miracidia}

\item{cercariae}{environmental miracidia}

\item{pars}{parameter set}
}
\description{
Save humans, environmental cercariae and mircacidia to a file along with the parameters used.
}
\examples{
\donttest{
 ## save_population_to_file(filename, humans, miracidia, cercariae, pars)
}
}
