% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{schistox_setup}
\alias{schistox_setup}
\title{Setup Schistoxpkg}
\usage{
schistox_setup(...)
}
\arguments{
\item{...}{Parameters are passed down to JuliaCall::julia_setup}
}
\description{
This function initializes Julia and the Schistoxpkg.jl package
The first time will be long since it includes precompilation
}
\examples{

\donttest{
## schistox_setup() is time-consuming and requires Julia

}

}
