% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{set_pars}
\alias{set_pars}
\title{Set the parameters to the desired values.}
\usage{
set_pars(
  N,
  time_step,
  N_communities,
  community_probs,
  community_contact_rate,
  density_dependent_fecundity,
  average_worm_lifespan,
  max_age,
  initial_worms,
  initial_miracidia,
  initial_miracidia_days,
  init_env_cercariae,
  worm_stages,
  contact_rate,
  max_fec_contact_rate_product,
  max_fecundity,
  age_contact_rates,
  ages_for_contacts,
  contact_rate_by_age_array,
  mda_adherence,
  mda_access,
  female_factor,
  male_factor,
  miracidia_maturity,
  birth_rate,
  human_cercariae_prop,
  predis_aggregation,
  cercariae_survival,
  miracidia_survival,
  death_prob_by_age,
  ages_for_death,
  r,
  vaccine_effectiveness,
  drug_effectiveness,
  spec_ages,
  ages_per_index,
  record_frequency,
  use_kato_katz,
  kato_katz_par,
  heavy_burden_threshold,
  rate_acquired_immunity,
  M0,
  human_larvae_maturity_time,
  egg_sample_size,
  input_ages,
  input_contact_rates,
  scenario
)
}
\arguments{
\item{N}{size of human population (Int)}

\item{time_step}{length of time in days to step forward each time (Float)}

\item{N_communities}{number of communities sharing the same environmental source (Int)}

\item{community_probs}{probability of an integer being in a given community. This governs
the population sizes in each community. This must be an array the length of
N_communities (Array(Float))}

\item{community_contact_rate}{contact rate with the environment for each of the communities.
This must be an array the length of N_communities (Array(Float))}

\item{density_dependent_fecundity}{decrease in egg production per worm due to high density
of worms (Float)}

\item{average_worm_lifespan}{average life expectancy of a worm (Float)}

\item{max_age}{maximum age of human in the population (Float)}

\item{initial_worms}{number of worms in each person to begin with. The actual number in
each person if chosen from a Poisson distirubiton with this mean (Int)}

\item{initial_miracidia}{initial number of miracidia larvae in the environment (Int)}

\item{initial_miracidia_days}{miracidia will age into cercariae larvae after a specified
number of days. This parameter will specify how many days of initial_miracidia we
will have already had in the environment (Int)}

\item{init_env_cercariae}{initial number of cercariae larvae in the environment (Int)}

\item{worm_stages}{how many age stages are there for the worms. Having 1 stage will give
Gamma distributed death ages, while more than 1 will result in Erlang distribution (Int)}

\item{contact_rate}{global contact rate for the uptake of larvae from the environment (Float)}

\item{max_fec_contact_rate_product}{the product of max fecundity and the contact rate in the population. Setting this
to a desired value is often a good way to ensure that the epidemic stays within a reasonable range, as when the max fecundity
increases, if the contact rate doesn't decrease appropriately, then the behaviour of the outbreak can be unrealistically
difficult to control. Contact rate can then be calculate by dividing this value by max fecundity or vice versa. (Float)}

\item{max_fecundity}{expected number of eggs from a single worm pair. The actual number
will be chosen from a distribution with this mean (Float)}

\item{age_contact_rates}{contact rate for chosen age groups(Array(Float))}

\item{ages_for_contacts}{age groups for specifying contact rates (Array(Int))}

\item{contact_rate_by_age_array}{array holding contact rate for each age (Array(Float))}

\item{mda_adherence}{proportion of people who adhere to the mda (Float)}

\item{mda_access}{proportion of people who have access to the mda (Float)}

\item{female_factor}{factor for altering the contact rate for females, if we choose to
have gender specific behaviour which affects contact rate (Float)}

\item{male_factor}{factor for altering the contact rate for males, if we choose to have
gender specific behaviour which affects contact rate (Float)}

\item{miracidia_maturity}{number of days after which miracidias will mature to cercariae (Int)}

\item{birth_rate}{rate of birth of humans (Float)}

\item{human_cercariae_prop}{proportion of cercariae which are able to infect humans (Float)}

\item{predis_aggregation}{aggregation for predisposition of individuals to uptake larvae.
This is chosen from a Gamma distribution with mean 1 for each individual and set for
life. If this is high, then the aggregation is low, meaning that most individuals
have roughly the same predisposition. If it is low, then the larvae become
concentrated in a few individuals. (Float)}

\item{cercariae_survival}{what proportion of cercariae survive from one time point
to the next (Float)}

\item{miracidia_survival}{what proportion of miracidia survive from one time point
to the next (Float)}

\item{death_prob_by_age}{for specified age range, what is the probability of dying
each year (Array(Float))}

\item{ages_for_death}{age ranges for death probabilities (Array(Float))}

\item{r}{aggregation parameter for negative binomially distributed egg production (Float)}

\item{vaccine_effectiveness}{efficacy of a vaccine if one is used(Float)}

\item{drug_effectiveness}{efficacy of a drug given during MDA(Float)}

\item{spec_ages}{number of individuals by age group which we specify if we want a
particular age distribution for the simulation (Array(Float))}

\item{ages_per_index}{how many different ages we include in the spec_ages parameter (Int)}

\item{record_frequency}{how often we should record the prevalence in the population
during simulation (Float)}

\item{use_kato_katz}{if 0, then don't use Kato-Katz (KK) for egg counts, if 1, use KK (Int)}

\item{kato_katz_par}{parameter for Gamma distribution if KK is used (Float)}

\item{heavy_burden_threshold}{number of eggs at which an individual is said to have a
heavy infection (Int)}

\item{rate_acquired_immunity}{rate at which immunity will be acquired for individuals.
This will be multiplied by the cumulative number of worms people have had throughout
their life to decide the level of immunity acquired (Float)}

\item{M0}{if a particular for of egg production is used, this parameter is required and is
a proxy for mean worm burden (Float)}

\item{human_larvae_maturity_time}{length of time in days after which a cercariae uptaken by
a human will mature into a worm (Int)}

\item{egg_sample_size}{the proportion of eggs which are sampled from each individual every time we
check their burden. This is between 0 and 1, with 1 meaning that all the eggs in the person are
sampled, hence giving the true burden. Typical value for a urine sample may be ~ 1/100}

\item{input_ages}{input ages for constructing contact array}

\item{input_contact_rates}{input rates for constructing contact array}

\item{scenario}{can be one of "low adult", "moderate adult", "high adult"}
}
\description{
Set the parameters to the desired values.
}
