% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{update_contact_rate}
\alias{update_contact_rate}
\title{Update the contact rates for individuals when they age}
\usage{
update_contact_rate(humans, pars)
}
\arguments{
\item{humans}{- human population}

\item{pars}{- parameters}
}
\description{
Update the contact rates for individuals when they age
}
