% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{update_parameters}
\alias{update_parameters}
\title{Update the parameter set used for the simulations This will take the existing parameters and update contact_rate, max_fecundity, predis_aggregation and density_dependent_fecundity}
\usage{
update_parameters(
  pars,
  contact_rate,
  max_fecundity,
  predis_aggregation,
  density_dependent_fecundity
)
}
\arguments{
\item{pars}{existing parameters}

\item{contact_rate}{updated contact rate}

\item{max_fecundity}{updated max fecundity}

\item{predis_aggregation}{updated aggregation of predisposition}

\item{density_dependent_fecundity}{updated density dependent fecundity}
}
\description{
Update the parameter set used for the simulations This will take the existing parameters and update contact_rate, max_fecundity, predis_aggregation and density_dependent_fecundity
}
