% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{update_parameters_individually}
\alias{update_parameters_individually}
\title{update a specific parameter}
\usage{
update_parameters_individually(pars, name, value)
}
\arguments{
\item{pars}{the parameters object}

\item{name}{the name of the parameter. This must match the name of a parameter defined in the set_pars function and be input as a string.}

\item{value}{the value to update the parameter to}
}
\description{
update a specific parameter
}
