% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{update_specified_parameters}
\alias{update_specified_parameters}
\title{Update a number of specified parameters at the same time}
\usage{
update_specified_parameters(pars, ...)
}
\arguments{
\item{pars}{the parameters object}

\item{...}{list of parameters and values for chosen parameters. Must be paired, with the name of the parameter in "" first, followed by the desired valeu.
For example, to update parameter N to 500 and max_fecunidty to 4  the function entry must look
like: update_specified_parameters(pars, "N", 500, "max_fecundity", 4)}
}
\description{
Update a number of specified parameters at the same time
}
