% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schistox.R
\name{add_to_mda}
\alias{add_to_mda}
\title{add to MDA schedule}
\usage{
add_to_mda(
  mda_info,
  mda_start_time,
  last_mda_time,
  regularity,
  drug_efficacy,
  pre_SAC_prop,
  SAC_prop,
  adult_prop
)
}
\arguments{
\item{mda_info}{current MDA information}

\item{mda_start_time}{start time for new MDAs}

\item{last_mda_time}{end time for new MDAs}

\item{regularity}{how regularly the MDA takes place in years (1 is once a year, 0.5 is twice a year)}

\item{drug_efficacy}{how effective the drug is}

\item{pre_SAC_prop}{what proportion of pre school age children receive MDA}

\item{SAC_prop}{what proportion of school age children receive MDA}

\item{adult_prop}{what proportion of adults receive MDA}
}
\description{
add to MDA schedule
}
